feat(collections): allow passing iterable to chunk
