+++ collections/chunk.ts
@@ -2,11 +2,11 @@
 // This module is browser compatible.
 
 /**
- * Splits the given array into chunks of the given size and returns them.
+ * Splits the given array into an array of chunks of the given size and returns them.
  *
- * @typeParam T Type of the elements in the input array.
+ * @typeParam T The type of the elements in the iterable.
  *
- * @param array The array to split into chunks.
+ * @param iterable The iterable to take elements from.
  * @param size The size of the chunks. This must be a positive integer.
  *
  * @returns An array of chunks of the given size.
@@ -37,20 +37,37 @@
  * );
  * ```
  */
-export function chunk<T>(array: readonly T[], size: number): T[][] {
+export function chunk<T>(
+  iterable: Iterable<T>,
+  size: number,
+): T[][] {
   if (size <= 0 || !Number.isInteger(size)) {
     throw new RangeError(
       `Expected size to be an integer greater than 0 but found ${size}`,
     );
   }
-
   const result: T[][] = [];
-  let index = 0;
 
-  while (index < array.length) {
-    result.push(array.slice(index, index + size));
-    index += size;
+  // Faster path
+  if (Array.isArray(iterable)) {
+    let index = 0;
+    while (index < iterable.length) {
+      result.push(iterable.slice(index, index + size));
+      index += size;
+    }
+    return result;
   }
 
+  let chunk: T[] = [];
+  for (const item of iterable) {
+    chunk.push(item);
+    if (chunk.length === size) {
+      result.push(chunk);
+      chunk = [];
+    }
+  }
+  if (chunk.length > 0) {
+    result.push(chunk);
+  }
   return result;
 }
